requirements.txt
absl-py==0.9.0
ai2thor==2.4.6
apache-beam==2.23.0
astor==0.8.1
avro-python3==1.9.2.1
aws-requests-auth==0.4.3
awscli==1.16.216
backcall==0.2.0
beautifulsoup4==4.9.3
blis==0.4.1
boto3==1.17.16
botocore==1.20.16
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.3.9
crcmod==1.7
cycler==0.10.0
cymem==2.0.3
Cython==0.29.20
decorator==4.4.2
dill==0.3.1.1
docopt==0.6.2
docutils==0.15.2
fastavro==0.23.6
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Dropzone==1.6.0
Flask-Reuploaded==1.1.0
ftfy==5.8
future==0.18.2
gast==0.2.2
gevent==20.12.1
google-api-core==1.21.0
google-api-python-client==2.4.0
google-auth==1.19.0
google-auth-httplib2==0.1.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.1
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.30.0
hdfs==2.5.8
httplib2==0.17.4
huggingface-hub==0.0.8
idna==2.10
importlib-metadata==1.7.0
ipdb==0.13.3
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lapjv==1.3.14
line-profiler==2.1.1
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
millify==0.1.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==2.0.0
msgpack==1.0.0
murmurhash==1.0.2
nlp==0.4.0
numpy==1.16.4
nvidia-pyindex==1.0.6
oauth2client==4.1.3
opencv-python==4.3.0.36
opt-einsum==3.2.1
orjson==3.4.6
packaging==20.9
parso==0.7.0
pbr==5.4.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
progressbar2==3.51.4
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==0.16.0
pyasn1==0.4.6
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
rsa==3.4.2
s3transfer==0.3.4
sacremoses==0.0.45
scikit-video==1.1.11
seaborn==0.10.1
sentencepiece==0.1.95
six==1.15.0
soupsieve==2.2.1
spacy==2.3.2
srsly==1.0.2
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.5
termcolor==1.1.0
thinc==7.4.1
tornado==6.0.4
traitlets==4.3.3
transformers==4.6.1
typing==3.7.4.1
typing-extensions==3.7.4.2
ujson==4.0.2
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.25.9
wasabi==0.7.1
Werkzeug==1.0.1
wrapt==1.12.1
xxhash==2.0.0
youtube-dl==2021.5.16
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0